# Conditionally enable Neon version of s2n-bignum implementations for montgomery multiplications