### Description of changes: 
This patch is the last patch splitted from #1108 for improving RSA signature speed for Graviton 2.
This conditionally enables the the Neon versions of verified primitives of s2n-bignum when `__ARM_NEON` is defined.
I checked this `__ARM_NEON` macro to be consistent with what BoringSSL does, according to [BUILDING.md](https://github.com/aws/aws-lc/blob/main/BUILDING.md).
The scalar versions of s2n-bignum primitives are already merged in b706d7e.

The performance improvements of RSA signatures on Graviton 2 are as follows (Unit: ops/sec).

|Bits|Operation|AWS-LC before b706d7e|AWS-LC after b706d7e|This patch|speedup vs. before b706d7e|
|:----|:----|:----|:----|:----|:----|
|2048|RSA sign|299.3|399|495.8|65.65%|
| |verify (fresh key)|10736.3|15491|18836.6|75.45%|
|3072|RSA sign|95.4|113.2|126.4|32.49%|
| |verify (fresh key)|4917.7|6001.7|6579.1|33.78%|
|4096|RSA sign|41.7|63.2|78.3|87.77%|
| |verify (fresh key)|2781.6|3451|3800.3|36.62%|

### Testing:
Tested via `bssl speed -filter RSA`

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license and the ISC license.